Created a seperate UserApi service and updated the websocket auth to support API keys
