+++ XiansAi.Server.Src/Features/UserApi/Auth/WebhookAuthenticationHandler.cs
@@ -4,33 +4,27 @@ using Shared.Auth;
 using Shared.Services;
 using System.Security.Claims;
 using System.Text.Encodings.Web;
-using XiansAi.Server.Shared.Services;
 
-namespace XiansAi.Server.Shared.Auth
+namespace Features.UserApi.Auth
 {
     public class WebhookAuthenticationHandler : AuthenticationHandler<AuthenticationSchemeOptions>
     {
         private readonly ITenantContext _tenantContext;
         private readonly ILogger<WebhookAuthenticationHandler> _logger;
-        private readonly IConfiguration _configuration;
-        private readonly IAuthorizationCacheService _authorizationCacheService;
         private readonly IApiKeyService _apiKeyService;
 
 
         public WebhookAuthenticationHandler(
             IOptionsMonitor<AuthenticationSchemeOptions> options,
             ILoggerFactory logger,
-            IAuthorizationCacheService authorizationCacheService,
             UrlEncoder encoder,
-            IConfiguration configuration,
             ITenantContext tenantContext,
             IApiKeyService apiKeyService)
             : base(options, logger, encoder)
         {
             _logger = logger.CreateLogger<WebhookAuthenticationHandler>();
             _tenantContext = tenantContext;
-            _configuration = configuration;
-            _authorizationCacheService = authorizationCacheService;
+
             _apiKeyService = apiKeyService;
         }
 
@@ -50,13 +44,7 @@ namespace XiansAi.Server.Shared.Auth
             }
 
             var tenantId = Request.Headers["X-Tenant-Id"].FirstOrDefault() ?? Request.Query["tenantId"].ToString();
-            var userAccessToken = Request.Headers["X-Secondary-Token"].FirstOrDefault() ?? null;
-            string? accessGuid = null;
-            if (!string.IsNullOrEmpty(userAccessToken))
-            {
-                accessGuid = await _authorizationCacheService.CacheAuthorization(userAccessToken);
-            }
-          
+
             if (_tenantContext != null)
             {
 
@@ -78,56 +66,67 @@ namespace XiansAi.Server.Shared.Auth
                 {
                     try
                     {
-                        // Get the apikey object
-                        var apiKey = await _apiKeyService.GetApiKeyByRawKeyAsync(accessToken, tenantId);
-                        // Check if the apikey section is null
-                        if (apiKey == null)
+                        if (accessToken.StartsWith("sk-Xnai-"))
                         {
-                            _logger.LogWarning("Webhook apikey not found");
-                            return AuthenticateResult.Fail("Invalid API key or Tenant ID");
-                        }
-                      
-                        // Verify if the provided tenant matches the expected tenant
-                        if (tenantId == apiKey.TenantId)
-                        {
+                            // Get the apikey object
+                            var apiKey = await _apiKeyService.GetApiKeyByRawKeyAsync(accessToken, tenantId);
+                            // Check if the apikey section is null
+                            if (apiKey == null)
+                            {
+                                _logger.LogWarning("Webhook apikey not found");
+                                return AuthenticateResult.Fail("Invalid API key or Tenant ID");
+                            }
 
-                            _tenantContext.LoggedInUser = apiKey.CreatedBy;
-                            _tenantContext.TenantId = apiKey.TenantId;
-                            _tenantContext.AuthorizedTenantIds = new[] { apiKey.TenantId };
+                            // Verify if the provided tenant matches the expected tenant
+                            if (tenantId == apiKey.TenantId)
+                            {
 
-                            var claims = new List<Claim>
+                                _tenantContext.LoggedInUser = apiKey.CreatedBy;
+                                _tenantContext.TenantId = apiKey.TenantId;
+                                _tenantContext.AuthorizedTenantIds = new[] { apiKey.TenantId };
+
+                                var claims = new List<Claim>
                             {
                                 new Claim(ClaimTypes.NameIdentifier, apiKey.CreatedBy),
                                 new Claim("TenantId", apiKey.TenantId)
                             };
-                            //if (!string.IsNullOrEmpty(accessGuid))
-                            //{
-                            //    claims.Add(new Claim("UserAccessGuid", accessGuid));
-                            //}
 
-                            var identity = new ClaimsIdentity(claims, Scheme.Name);
-                            var principal = new ClaimsPrincipal(identity);
-                            var ticket = new AuthenticationTicket(principal, Scheme.Name);
-                            _logger.LogInformation("Successfully authenticated Webhook connection: User={UserId}, Tenant={TenantId}", apiKey.CreatedBy, tenantId);
+                                var identity = new ClaimsIdentity(claims, Scheme.Name);
+                                var principal = new ClaimsPrincipal(identity);
+                                var ticket = new AuthenticationTicket(principal, Scheme.Name);
+                                _logger.LogInformation("Successfully authenticated Webhook connection: User={UserId}, Tenant={TenantId}", apiKey.CreatedBy, tenantId);
 
-                            return AuthenticateResult.Success(ticket);
+                                return AuthenticateResult.Success(ticket);
+                            }
+                            else
+                            {
+                                _logger.LogWarning("Invalid TenantID {TenantId}", tenantId);
+                                return AuthenticateResult.Fail("Access denied Invalid TenantID");
+                            }
+                        }
+                        else if (accessToken.Count(c => c == '.') == 2)
+                        {
+                            // Treat as JWT
+                            // TODO: Add your JWT validation logic here
+                            _logger.LogWarning("JWT authentication not implemented");
+                            return AuthenticateResult.Fail("JWT authentication not implemented");
                         }
                         else
                         {
-                            _logger.LogWarning("Access token does not match the expected secret for tenant {TenantId}", tenantId);
-                            return AuthenticateResult.Fail("Access token does not match the expected secret for tenant");
+                            _logger.LogWarning("Invalid token format");
+                            return AuthenticateResult.Fail("Invalid token format");
                         }
                     }
                     catch (Exception ex)
                     {
-                        _logger.LogError(ex, "Error processing access token for SignalR connection");
-                        return AuthenticateResult.Fail("Error processing access token for SignalR connection");
+                        _logger.LogError(ex, "Error processing access token for Webhook connection");
+                        return AuthenticateResult.Fail("Error processing access token for Webhook connection");
                     }
                 }
                 else
                 {
-                    _logger.LogWarning("No access token found for SignalR connection");
-                    return AuthenticateResult.Fail("No access token found for SignalR connection");
+                    _logger.LogWarning("No ApiKey found for Webhook trigger");
+                    return AuthenticateResult.Fail("No ApiKey found for Webhook trigger");
                 }
             }
             else
