+++ XiansAi.Server.Src/Features/WebApi/Configuration/WebApiConfiguration.cs
@@ -2,7 +2,6 @@ using Features.WebApi.Auth;
 using Features.WebApi.Endpoints;
 using Features.WebApi.Repositories;
 using Features.WebApi.Services;
-using Microsoft.AspNetCore.Authorization;
 using XiansAi.Server.Features.WebApi.Endpoints;
 using XiansAi.Server.Features.WebApi.Repositories;
 using XiansAi.Server.Features.WebApi.Services;
@@ -21,7 +20,6 @@ public static class WebApiConfiguration
         builder.Services.AddScoped<IWorkflowCancelService, WorkflowCancelService>();
         builder.Services.AddScoped<ILogsService, LogsService>();
         builder.Services.AddScoped<ITenantService, TenantService>();
-        //builder.Services.AddScoped<IWebhookService, WebhookService>();
         builder.Services.AddScoped<IActivitiesService, ActivitiesService>();
         builder.Services.AddScoped<IMessagingService, MessagingService>();
         builder.Services.AddScoped<IAuditingService, AuditingService>();
@@ -33,7 +31,6 @@ public static class WebApiConfiguration
         // Register repositories
         builder.Services.AddScoped<ILogRepository, LogRepository>();
         builder.Services.AddScoped<IActivityRepository, ActivityRepository>();
-        //builder.Services.AddScoped<IWebhookRepository, WebhookRepository>();
         builder.Services.AddScoped<ITenantRepository, TenantRepository>();
         builder.Services.AddScoped<IUserRoleRepository, UserRoleRepository>();
 
@@ -48,7 +45,7 @@ public static class WebApiConfiguration
         LogsEndpoints.MapLogsEndpoints(app);
         SettingsEndpoints.MapSettingsEndpoints(app);
         TenantEndpoints.MapTenantEndpoints(app);
-        //WebhookEndpoints.MapWebhookEndpoints(app);
+        WebhookEndpoints.MapWebhookEndpoints(app);
         PublicEndpoints.MapPublicEndpoints(app);
         MessagingEndpoints.MapMessagingEndpoints(app);
         AuditingEndpoints.MapAuditingEndpoints(app);
